fix: correct MockDEX ABI/directions, swap TradeApprovalModal to ContractExecuteTransaction, fix TS errors
MockDEX fixes (tradeExecutor.ts + TradeApprovalModal.tsx):
- Replace non-existent sellHBARforUSDT/buyHBARwithUSDT with the single
  executeSwap(agentId, direction, amountIn, minAmountOut, hcsSeq, topicId)
  function that the deployed contract actually exposes
- Direction strings corrected: HBAR_TO_USDT → HBAR_TO_USDC,
  USDT_TO_HBAR → USDC_TO_HBAR (matching on-chain revert message)
- MOCK_DEX_ABI in both files updated to match deployed contract

TradeApprovalModal.tsx — fix _makeTransactionBody TypeError:
- DAppSigner.signTransaction expects a Hedera SDK Transaction object;
  passing a raw EVM tx via hashpackEthers bridge causes the error
- Remove getHashPackEthersSigner entirely; replace executeSwap write
  call with ContractExecuteTransaction + ContractFunctionParameters
  (freezeWithSigner → executeWithSigner, same pattern as deployment)
- getSwapQuote read call kept as ethers.JsonRpcProvider (no signing needed)

TypeScript fixes:
- agentStore.ts: add explicit optional typed fields (agentId, indicators,
  risk) to AgentConfig — prevents [key:string]:unknown widening all
  property accesses to {} in older TS, fixing 24 errors in create/page.tsx
- create/page.tsx: BigInt(Math.round(...)) → Math.round(...) for
  Hbar.fromTinybars() which does not accept native bigint
- apps/api/tsconfig.json: rootDir "src" → "../.." and add packages to
  include array to resolve TS6059 for @tradeagent/hedera imports

Config updates:
- Default GEMINI_MODEL updated to gemini-2.5-flash in env.ts + .env.example

Docs:
- bugs.md: correct Bug 14 (was documenting wrong fix), add Bugs 16-18
- enhancement.md: Enhancement 2 updated to document ContractExecuteTransaction
  now covers both deploy and manual trade, with read/write split table

Made-with: Cursor

+++ apps/web/src/app/create/page.tsx
@@ -94,7 +94,7 @@ function FundAgentModal({ agentId, agentAccountId, signer, accountId, onComplete
     setFunding(true);
     setError(null);
     try {
-      const tinybars = BigInt(Math.round(hbar * 1e8));
+      const tinybars = Math.round(hbar * 1e8);
       // Build a simple HBAR transfer from user → agentAccountId
       const transferTx = await new TransferTransaction()
         .addHbarTransfer(AccountId.fromString(accountId), Hbar.fromTinybars(-tinybars))
